Converted classes to records
