+++ tradingcards-api/src/main/java/net/tinetwork/tradingcards/api/model/chance/EmptyChance.java
@@ -1,4 +1,4 @@
-package net.tinetwork.tradingcards.api.model;
+package net.tinetwork.tradingcards.api.model.chance;
 
 public class EmptyChance extends Chance{
     public EmptyChance() {
